feat: Implement Book rating System
Added react-stars to implement a rating system for books. So far, the API seems to only allow to read the ratings of the book. The user is able to rate the book, and an api put request is fired to update the book's average rating according to the user's new rating, but it doesn't seem to allow the user to actually change the record in the database for some reason. I'm still investigating this, but so far, we have a read-only star rating system.